+++ spec/rails_helper.rb
@@ -1,40 +1,18 @@
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 require 'spec_helper'
-require 'support/factory_bot'
-require_relative 'support/controller_macros'
-require_relative 'support/fake_objects'
-require 'webmock/rspec'
-require_relative './features/helpers'
-WebMock.disable_net_connect!(allow_localhost: true)
 ENV['RAILS_ENV'] ||= 'test'
 
 require File.expand_path('../config/environment', __dir__)
-include Warden::Test::Helpers
 
 # Prevent database truncation if the environment is production
-abort("The Rails environment is running in production mode!") if Rails.env.production?
+abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
-require "view_component/test_helpers"
+require 'view_component/test_helpers'
 
 require 'axe/rspec'
-require 'capybara/rspec'
-require 'capybara/apparition'
 # Add additional requires below this line. Rails is not loaded until this point!
 
-# Requires supporting ruby files with custom matchers and macros, etc, in
-# spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
-# run as spec files by default. This means that files in spec/support that end
-# in _spec.rb will both be required and run as specs, causing the specs to be
-# run twice. It is recommended that you do not name files matching this glob to
-# end with _spec.rb. You can configure this pattern with the --pattern
-# option on the command line or in ~/.rspec, .rspec or `.rspec-local`.
-#
-# The following line is provided for convenience purposes. It has the downside
-# of increasing the boot-up time by auto-requiring all files in the support
-# directory. Alternatively, in the individual `*_spec.rb` files, manually
-# require only the support files necessary.
-#
-# Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
+Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each do |f| require f end
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.
@@ -44,17 +22,8 @@ rescue ActiveRecord::PendingMigrationError => e
   puts e.to_s.strip
   exit 1
 end
+
 RSpec.configure do |config|
-
-  # Exclude running accessibility tests by default
-  config.filter_run_excluding :accessibility => true
-
-  # Set the Capybara driver when running accessibility tests using
-  # bundle exec rspec --tag accessibility
-  if config.filter_manager.inclusions.rules.include?(:accessibility)
-    Capybara.current_driver = :apparition
-  end
-
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"
 
@@ -85,20 +54,10 @@ RSpec.configure do |config|
 
   # Include access to Devise helper methods, such as sign_in
   config.include Devise::Test::ControllerHelpers, type: :controller
-
   config.extend ControllerMacros, type: :controller
   config.extend FakeObjects
-  config.before(:suite) do
-    DatabaseCleaner.strategy = :transaction
-    DatabaseCleaner.clean_with(:truncation)
-  end
 
-  config.around(:each) do |example|
-    DatabaseCleaner.cleaning do
-      example.run
-    end
-  end
-
-  config.include Helpers
+  config.include FeatureHelpers, type: :feature
   config.include ViewComponent::TestHelpers, type: :component
+  config.include Warden::Test::Helpers
 end
