+++ spec/support/controller_macros.rb
@@ -1,7 +1,7 @@
 module ControllerMacros
   def login_user
     before(:each) do
-      @request.env["devise.mapping"] = Devise.mappings[:user]
+      @request.env['devise.mapping'] = Devise.mappings[:user]
       user = create(:user)
       sign_in user
     end
