+++ spec/support/fake_objects.rb
@@ -1,22 +1,19 @@
 module FakeObjects
-
   # Mock class for HTTP responses
   class FakeHttpResponse
-
-    attr_reader :status, :body
+    attr_accessor :status
     attr_writer :status, :body
-  
+
     def initialize(status, body)
       self.status = status
       self.body = body
     end
-  
   end
 
   # Mock class for the NotifyMailer class
   class FakeNotifyMailer
     def deliver_later
-      return 'deliver_later'
+      'deliver_later'
     end
   end
 
@@ -24,12 +21,12 @@ module FakeObjects
   class FakeFile
     class FakeBlob
       def filename
-        return 'filename.txt'
+        'filename.txt'
       end
     end
+
     def blob
-      return FakeBlob.new
+      FakeBlob.new
     end
   end
-
 end
